+++ R/ggplot2-utils.R
@@ -1,9 +1,46 @@
+# Theme -------------------------------------------------------------------
+
 # generic theme that makes a ggplot2 look like the old lattice plots
-theme_openair <- function(key.position) {
+theme_openair <- function(key.position, extra.args) {
+  fontsize <- set_extra_fontsize(extra.args)
+
   if (user_has_set_theme()) {
     return(
+      list(
+        ggplot2::theme(
+          legend.position = key.position,
+          legend.ticks.length = structure(
+            if (key.position %in% c("bottom", "right")) {
+              c(-0.2, 0)
+            } else {
+              c(0, -0.2)
+            },
+            class = "rel"
+          )
+        ),
+        fontsize
+      )
+    )
+  }
+
+  list(
+    ggplot2::theme_bw() +
       ggplot2::theme(
+        strip.background = ggplot2::element_rect(fill = "white"),
+        panel.spacing = ggplot2::rel(2.5),
         legend.position = key.position,
+        legend.background = ggplot2::element_blank(),
+        plot.title = ggplot2::element_text(hjust = 0.5, face = "bold"),
+        plot.subtitle = ggplot2::element_text(hjust = 0.5),
+        plot.caption = ggplot2::element_text(hjust = 0.5, face = "bold"),
+        plot.margin = ggplot2::rel(4),
+        legend.frame = ggplot2::element_rect(
+          fill = NA,
+          color = "black",
+          linewidth = 0.25
+        ),
+        legend.title = ggplot2::element_text(hjust = 0.5),
+        legend.ticks = ggplot2::element_line(),
         legend.ticks.length = structure(
           if (key.position %in% c("bottom", "right")) {
             c(-0.2, 0)
@@ -11,45 +48,29 @@ theme_openair <- function(key.position) {
             c(0, -0.2)
           },
           class = "rel"
-        )
-      )
-    )
-  }
-
-  ggplot2::theme_bw() +
-    ggplot2::theme(
-      strip.background = ggplot2::element_rect(fill = "white"),
-      panel.spacing = ggplot2::rel(2.5),
-      legend.position = key.position,
-      legend.background = ggplot2::element_blank(),
-      plot.title = ggplot2::element_text(hjust = 0.5, face = "bold"),
-      plot.subtitle = ggplot2::element_text(hjust = 0.5),
-      plot.caption = ggplot2::element_text(hjust = 0.5, face = "bold"),
-      plot.margin = ggplot2::rel(4),
-      legend.frame = ggplot2::element_rect(
-        fill = NA,
-        color = "black",
-        linewidth = 0.25
+        ),
+        plot.background = ggplot2::element_rect(
+          fill = "transparent",
+          color = NA
+        ),
+        strip.text = ggplot2::element_text(margin = ggplot2::margin_auto(1))
       ),
-      legend.title = ggplot2::element_text(hjust = 0.5),
-      legend.ticks = ggplot2::element_line(),
-      legend.ticks.length = structure(
-        if (key.position %in% c("bottom", "right")) c(-0.2, 0) else c(0, -0.2),
-        class = "rel"
-      ),
-      plot.background = ggplot2::element_rect(fill = "transparent", color = NA),
-      strip.text = ggplot2::element_text(margin = ggplot2::margin_auto(1))
-    )
+    fontsize
+  )
 }
 
 # theme for radial plots
-theme_openair_radial <- function(key.position, panel.ontop = FALSE) {
+theme_openair_radial <- function(
+  key.position,
+  extra.args,
+  panel.ontop = FALSE
+) {
   if (user_has_set_theme()) {
-    return(theme_openair(key.position))
+    return(theme_openair(key.position, extra.args))
   }
 
   list(
-    theme_openair(key.position),
+    theme_openair(key.position, extra.args),
     ggplot2::theme(
       axis.text = ggplot2::element_text(color = "black"),
       axis.ticks.theta = ggplot2::element_blank(),
@@ -82,13 +103,13 @@ theme_openair_radial <- function(key.position, panel.ontop = FALSE) {
 }
 
 # adapted theme_openair with a (by default) blue dashed gridline
-theme_openair_sf <- function(key.position, grid.col) {
+theme_openair_sf <- function(key.position, extra.args, grid.col) {
   if (user_has_set_theme()) {
-    return(theme_openair(key.position))
+    return(theme_openair(key.position, extra.args))
   }
 
   list(
-    theme_openair(key.position),
+    theme_openair(key.position, extra.args),
     ggplot2::theme(
       panel.grid = ggplot2::element_line(
         colour = grid.col,
@@ -103,11 +124,7 @@ theme_openair_sf <- function(key.position, grid.col) {
   )
 }
 
-user_has_set_theme <- function() {
-  !identical(ggplot2::get_theme(), ggplot2::theme_gray())
-}
-
-# take the extra.args and set a different global fontsize, if present
+# handle "fontsize" arg - used in theme_openair
 set_extra_fontsize <- function(extra.args) {
   if ("fontsize" %in% names(extra.args)) {
     list(
@@ -119,6 +136,13 @@ set_extra_fontsize <- function(extra.args) {
     list()
   }
 }
+
+# check if a theme is set
+user_has_set_theme <- function() {
+  !identical(ggplot2::get_theme(), ggplot2::theme_gray())
+}
+
+# Facet -------------------------------------------------------------------
 
 # work out the faceting strategy
 get_facet <- function(
@@ -185,6 +209,7 @@ get_facet <- function(
   fun
 }
 
+# convert `lattice` 'relation' to `ggplot2`
 relation_to_facet_scales <- function(x.relation, y.relation) {
   x.relation <- x.relation == "free"
   y.relation <- y.relation == "free"
@@ -195,6 +220,9 @@ relation_to_facet_scales <- function(x.relation, y.relation) {
     !x.relation && !y.relation ~ "fixed"
   )
 }
+
+
+# Scales ------------------------------------------------------------------
 
 # ggplot2 scale for radial plots
 scale_x_compass <- function(
@@ -216,6 +244,9 @@ scale_x_compass <- function(
     )
   )
 }
+
+
+# Layers ------------------------------------------------------------------
 
 annotate_compass_points <- function(size, labels = c("N", "E", "S", "W")) {
   list(
@@ -256,27 +287,6 @@ annotate_compass_points <- function(size, labels = c("N", "E", "S", "W")) {
       fontface = "bold"
     )
   )
-}
-
-# Recycle helper similar to lattice behaviour
-recycle_to_length <- function(x, n, expect1 = FALSE) {
-  if (length(x) == n) {
-    return(x)
-  }
-  if (length(x) == 1) {
-    return(rep(x, n))
-  }
-
-  if (expect1) {
-    cli::cli_abort(
-      "Length mismatch: argument must be length 1 or same length as 'h'/'v'"
-    )
-  } else {
-    while (length(x) < n) {
-      x <- c(x, x)
-    }
-    x <- x[seq_len(x)]
-  }
 }
 
 # Convert lattice-style ref.y list to ggplot2 geom_hline layers
@@ -340,6 +350,3 @@ gg_ref_x <- function(ref.x) {
     lwd
   )
 }
-
-# `%||%` for convenience
-`%||%` <- function(a, b) if (!is.null(a)) a else b