+++ R/percentileRose.R
@@ -521,8 +521,11 @@ percentileRose <- function(
       limits = c(0, ifelse(is.null(intervals), NA, max(intervals))),
       breaks = intervals %||% scales::breaks_pretty()
     ) +
-    theme_openair_radial(key.position = key.position, panel.ontop = TRUE) +
-    set_extra_fontsize(extra.args) +
+    theme_openair_radial(
+      key.position = key.position,
+      extra.args = extra.args,
+      panel.ontop = TRUE
+    ) +
     ggplot2::labs(
       x = extra.args$xlab,
       y = extra.args$ylab,