+++ R/polarCluster.R
@@ -320,8 +320,11 @@ polarCluster <-
         ),
         drop = FALSE
       ) +
-      theme_openair_radial(key.position %||% "right", panel.ontop = TRUE) +
-      set_extra_fontsize(extra.args) +
+      theme_openair_radial(
+        key.position %||% "right",
+        extra.args = extra.args,
+        panel.ontop = TRUE
+      ) +
       annotate_compass_points(
         size = ifelse(
           extra.args$annotate %||% TRUE,