+++ R/polarPlot.R
@@ -982,8 +982,11 @@ polarPlot <-
           )
         }
       } +
-      theme_openair_radial(key.position, panel.ontop = TRUE) +
-      set_extra_fontsize(extra.args) +
+      theme_openair_radial(
+        key.position,
+        extra.args = extra.args,
+        panel.ontop = TRUE
+      ) +
       annotate_compass_points(
         size = ifelse(
           extra.args$annotate,