+++ R/trajPlot.R
@@ -290,8 +290,11 @@ trajPlot <- function(
 
   # base plot & themes
   thePlot <- ggplot2::ggplot() +
-    theme_openair_sf(key.position, grid.col = grid.col) +
-    set_extra_fontsize(extra.args) +
+    theme_openair_sf(
+      key.position,
+      extra.args = extra.args,
+      grid.col = grid.col
+    ) +
     get_facet(
       type,
       extra.args,